fix(): Correct visualization
