+++ src/main/java/com/techCamp/backend/components/JSONArraySerializer.java
@@ -1,13 +1,8 @@
 package com.techCamp.backend.components;
-
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializerProvider;
-import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-
 import org.json.JSONArray;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 
@@ -26,7 +21,6 @@ public class JSONArraySerializer extends StdSerializer<JSONArray>{
 
     @Override
     public void serialize(JSONArray value, JsonGenerator jgen, SerializerProvider provider) throws IOException {
-        System.out.println("*************************************+");
         jgen.writeRawValue(value.toString());
     }
 }
