+++ src/main/java/com/techCamp/backend/configuration/MongoConfig.java
@@ -5,6 +5,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
+import org.codehaus.plexus.component.annotations.Component;
 import org.json.JSONArray;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -14,7 +15,6 @@ import org.springframework.data.mongodb.config.AbstractMongoClientConfiguration;
 import org.springframework.data.mongodb.core.convert.DbRefResolver;
 import org.springframework.data.mongodb.core.convert.DefaultDbRefResolver;
 import org.springframework.data.mongodb.core.convert.MappingMongoConverter;
-import org.springframework.data.mongodb.core.convert.MongoConverter;
 import org.springframework.data.mongodb.core.convert.MongoCustomConversions;
 import org.springframework.data.mongodb.core.mapping.MongoMappingContext;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
@@ -23,6 +23,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
+import com.techCamp.backend.components.JSONArrayDeserializer;
 import com.techCamp.backend.components.JSONArrayReadConverter;
 import com.techCamp.backend.components.JSONArraySerializer;
 import com.techCamp.backend.components.JSONObjectReadConverter;
@@ -56,15 +57,6 @@ public class MongoConfig extends AbstractMongoClientConfiguration {
         converters.add(new JSONObjectReadConverter());
         MongoCustomConversions conversions=new MongoCustomConversions(converters);
         return conversions;
-    }
-
-    @Bean
-    public ObjectMapper getObjectMapper() {
-        var objectMapper = new ObjectMapper();
-        SimpleModule module = new SimpleModule();
-        module.addSerializer(JSONArray.class, new JSONArraySerializer());
-        objectMapper.registerModule(module);
-        return objectMapper;
     }
 
     @Bean
