+++ src/main/java/com/techCamp/backend/configuration/WebConfiguration.java
@@ -1,13 +1,23 @@
 package com.techCamp.backend.configuration;
 
+import java.util.List;
+
+import org.json.JSONArray;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.module.SimpleModule;
+import com.techCamp.backend.components.JSONArraySerializer;
 
 @Configuration
 @EnableWebMvc
@@ -27,4 +37,26 @@ public class WebConfiguration implements WebMvcConfigurer {
         source.registerCorsConfiguration("/**", corsConfiguration);
         return source;
     }
+     @Bean
+    public MappingJackson2HttpMessageConverter jacksonMessageConverter(ObjectMapper objectMapper) {
+        MappingJackson2HttpMessageConverter messageConverter = new MappingJackson2HttpMessageConverter();
+        messageConverter.setObjectMapper(objectMapper);
+        return messageConverter;
+    }
+
+    @Override
+    public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
+        converters.add(jacksonMessageConverter(getObjectMapper()));
+    }
+
+    @Bean
+    @Primary
+    public ObjectMapper getObjectMapper() {
+        var objectMapper = new ObjectMapper();
+        SimpleModule module = new SimpleModule();
+        module.addSerializer(JSONArray.class, new JSONArraySerializer());
+        objectMapper.registerModule(module);
+        return objectMapper;
+    }
+
 }
