+++ src/main/java/com/techCamp/backend/dto/CreateTableDTO.java
@@ -21,7 +21,6 @@ public class CreateTableDTO {
             JSONObject jsonObject = new JSONObject(dataMap);
             jsonArray.put(jsonObject);
         }
-        System.out.println(jsonArray.toString());
         return jsonArray;
     }
 }
